Corregido problema que solo se creaba una sola instancia por pantalla y agregada funcionalidad de grabar de las configuraciones

+++ workspace/recursos/src/main/resources/sql/create_impuesto_detalle.sql
@@ -15,7 +15,7 @@ create table IMPUESTO_DETALLE
     CODIGO integer,
     NOMBRE varchar(100),
     PORCENTAJE decimal,
-    TARIFA decimal,
+    TARIFA integer,
     DESCRIPCION varchar(100),
     FECHA_INICIO date,
     FECHA_FIN date,
